backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
curl_cffi==0.11.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
frozendict==2.4.6
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jiter==0.10.0
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
matplotlib-inline==0.1.7
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.82.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.50
protobuf==6.31.0
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yfinance==0.2.61
zipp==3.21.0
